# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: General Supplies Procurement Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SUPPLIERS ───────────────────────────────────────────────────────────────
...
  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List all inquiries
      responses:
        "200":
          description: List of inquiries with items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InquiryWithItems"
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Create a new inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /inquiries/{id}:
    get:
      operationId: getInquiry
      tags: [inquiries]
      summary: Get an inquiry by ID (with items)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inquiry with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryWithItems"
        "404":
          description: Not found
    patch:
      operationId: updateInquiry
      tags: [inquiries]
      summary: Update an inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    delete:
      operationId: deleteInquiry
      tags: [inquiries]
      summary: Delete an inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /inquiries/{id}/items:
    post:
      operationId: addInquiryItem
      tags: [inquiries]
      summary: Add an item to an inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryItem"
  /inquiries/{id}/items/{itemId}:
    patch:
      operationId: updateInquiryItem
      tags: [inquiries]
      summary: Update an inquiry item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InquiryItem"
    delete:
      operationId: deleteInquiryItem
      tags: [inquiries]
      summary: Delete an inquiry item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── QUOTATIONS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string

    # ─── SUPPLIER ──────────────────────────────────────────────────────────────
...
    Inquiry:
      type: object
      required: [id, customerId, title, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, in_progress, quoted, closed]
        createdAt:
          type: string
          format: date-time
    InquiryItem:
      type: object
      required: [id, inquiryId, description, quantity]
      properties:
        id:
          type: integer
        inquiryId:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    InquiryWithItems:
      type: object
      required: [id, customerId, title, status, createdAt, items]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, in_progress, quoted, closed]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/InquiryItem"
    InquiryInput:
      type: object
      required: [customerId, title]
      properties:
        customerId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [new, in_progress, quoted, closed]
    InquiryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [new, in_progress, quoted, closed]
    InquiryItemInput:
      type: object
      required: [description, quantity]
      properties:
        description:
          type: string
          minLength: 1
        quantity:
          type: number
        unit:
          type: string
        notes:
          type: string
    InquiryItemUpdate:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        quantity:
          type: number
        unit:
          type: string
        notes:
          type: string

    # ─── QUOTATION ─────────────────────────────────────────────────────────────
...