# lib/api-zod/src/generated/types/inquiryItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * General Supplies Procurement Management System
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryItemUpdate {
  /** @minLength 1 */
  description?: string;
  quantity?: number;
  unit?: string;
  notes?: string;
}